Extract the common types
