+++ src/classpath/assets/utils.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { ProjectType } from "../types";
+import { ProjectType } from "../../utils/webview";
 
 export const WEBVIEW_ID = "java.classpathConfiguration";
