+++ src/java-runtime/types.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+import { ProjectType } from "../utils/webview";
+
 export interface JavaRuntimeEntry {
   name: string;
   fspath: string;
@@ -22,18 +24,4 @@ export interface ProjectRuntimeEntry {
   runtimePath: string;
   sourceLevel: string;
   projectType: ProjectType;
-}
-
-export enum ProjectType {
-  Default = "Default project",
-  UnmanagedFolder = "Unmanaged folder",
-  Maven = "Maven",
-  Gradle = "Gradle",
-  Others = "Others",
-}
-
-export enum NatureId {
-  Maven = "org.eclipse.m2e.core.maven2Nature",
-  Gradle = "org.eclipse.buildship.core.gradleprojectnature",
-  Java = "org.eclipse.jdt.core.javanature",
 }
